// src/model/contact-dto.ts
export class ContactDto {
    id: string;
    salutationType: string; // MR, MRS,...
    nameInitial: string;
    firstName: string;
    lastName: string;
    name: string;
    fullName: string;
    avatarUrl: string;
    email: string;
    phoneNumber: string;
    homePhone: string;
    fax: string;
    website: string;
    address: string;
    createdDate: Date;
    modifiedDate: Date;
}

export class ContactDetailDto {

}